Refactor: 패키지명 수정 - 패키지명 수정 ( step3 -> lottoauto)
