+++ src/main/java/lottoauto/io/PrintUtils.java
@@ -1,0 +1,7 @@
+package lottoauto.io;
+
+public abstract class PrintUtils {
+    public static void println(String message){
+        System.out.println(message);
+    }
+}
