--- src/main/java/step3/domain/LottoNumber.java
@@ -1,37 +1,0 @@
-package step3.domain;
-
-import java.util.Objects;
-
-public class LottoNumber {
-    public static final int MIN_NUMBER = 1;
-    public static final int MAX_NUMBER = 45;
-    private int number;
-    public LottoNumber(int number) {
-        validateRange(number);
-        this.number = number;
-    }
-
-    private void validateRange(int number) {
-        if(number < MIN_NUMBER || number > MAX_NUMBER){
-            throw new IllegalArgumentException("로또번호의 범위(1~45)에 맞지 않습니다.");
-        }
-    }
-
-    @Override
-    public boolean equals(Object o) {
-        if (this == o) return true;
-        if (o == null || getClass() != o.getClass()) return false;
-        LottoNumber number1 = (LottoNumber) o;
-        return number == number1.number;
-    }
-
-    @Override
-    public int hashCode() {
-        return Objects.hash(number);
-    }
-
-    @Override
-    public String toString() {
-        return String.valueOf(number);
-    }
-}
