--- src/main/java/step3/io/InputUtils.java
@@ -1,11 +1,0 @@
-package step3.io;
-
-import java.util.Scanner;
-
-public abstract class InputUtils {
-    private static final Scanner scanner = new Scanner(System.in);
-
-    public static String readConsole(){
-        return scanner.nextLine();
-    }
-}
