--- src/main/java/step3/io/PrintUtils.java
@@ -1,7 +1,0 @@
-package step3.io;
-
-public abstract class PrintUtils {
-    public static void println(String message){
-        System.out.println(message);
-    }
-}
