--- src/main/java/step3/strategy/LottoNumberStrategy.java
@@ -1,8 +1,0 @@
-package step3.strategy;
-
-import java.util.List;
-
-public interface LottoNumberStrategy {
-
-    List<Integer> generateNumbers();
-}
