--- src/main/java/step3/view/InputView.java
@@ -1,12 +1,0 @@
-package step3.view;
-
-import static step3.io.InputUtils.*;
-import static step3.io.PrintUtils.println;
-
-public abstract class InputView {
-
-    public static String payLotto() {
-        println("구입금액을 입력하세요.");
-        return readConsole();
-    }
-}
